+++ anvil_consortium_manager/views.py
@@ -5,7 +5,7 @@ from django.contrib.auth.mixins import LoginRequiredMixin
 from django.contrib.sites.shortcuts import get_current_site
 from django.core.exceptions import ImproperlyConfigured
 from django.db import transaction
-from django.forms import Form, inlineformset_factory
+from django.forms import Form, HiddenInput, inlineformset_factory
 from django.http import Http404, HttpResponseRedirect
 from django.urls import reverse
 from django.utils import timezone
@@ -1779,6 +1779,199 @@ class WorkspaceGroupSharingCreate(
             return self.render_to_response(self.get_context_data(form=form))
         # The object is saved by the super's form_valid method.
         return super().form_valid(form)
+
+
+class WorkspaceGroupSharingCreateByWorkspace(WorkspaceGroupSharingCreate):
+    """View to create a new WorkspaceGroupSharing object for the workspace specified in the url."""
+
+    model = models.WorkspaceGroupSharing
+    form_class = forms.WorkspaceGroupSharingForm
+    template_name = (
+        "anvil_consortium_manager/workspacegroupsharing_form_byworkspace.html"
+    )
+    success_msg = "Successfully shared Workspace with Group."
+    """Message to display when the WorkspaceGroupSharing object was successfully created in the app and on AnVIL."""
+
+    message_group_not_found = "Managed Group not found on AnVIL."
+    """Message to display when the ManagedGroup was not found on AnVIL."""
+
+    message_already_exists = (
+        "This workspace has already been shared with this managed group."
+    )
+
+    def get_workspace(self):
+        try:
+            billing_project_slug = self.kwargs["billing_project_slug"]
+            workspace_slug = self.kwargs["workspace_slug"]
+            workspace = models.Workspace.objects.get(
+                billing_project__name=billing_project_slug, name=workspace_slug
+            )
+        except models.Workspace.DoesNotExist:
+            raise Http404("Workspace not found.")
+        return workspace
+
+    def get(self, request, *args, **kwargs):
+        self.workspace = self.get_workspace()
+        return super().get(request, *args, **kwargs)
+
+    def post(self, request, *args, **kwargs):
+        self.workspace = self.get_workspace()
+        return super().post(request, *args, **kwargs)
+
+    def get_initial(self):
+        initial = super().get_initial()
+        initial["workspace"] = self.workspace
+        return initial
+
+    def get_form(self, **kwargs):
+        """Get the form and set the inputs to use a hidden widget."""
+        form = super().get_form(**kwargs)
+        form.fields["workspace"].widget = HiddenInput()
+        return form
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context["workspace"] = self.workspace
+        return context
+
+    def get_success_url(self):
+        return self.object.get_absolute_url()
+
+
+class WorkspaceGroupSharingCreateByGroup(WorkspaceGroupSharingCreate):
+    """View to create a new WorkspaceGroupSharing object for the group specified in the url."""
+
+    model = models.WorkspaceGroupSharing
+    form_class = forms.WorkspaceGroupSharingForm
+    template_name = "anvil_consortium_manager/workspacegroupsharing_form_bygroup.html"
+    success_msg = "Successfully shared Workspace with Group."
+    """Message to display when the WorkspaceGroupSharing object was successfully created in the app and on AnVIL."""
+
+    message_group_not_found = "Managed Group not found on AnVIL."
+    """Message to display when the ManagedGroup was not found on AnVIL."""
+
+    message_already_exists = (
+        "This workspace has already been shared with this managed group."
+    )
+
+    def get_group(self):
+        try:
+            group_slug = self.kwargs["group_slug"]
+            group = models.ManagedGroup.objects.get(name=group_slug)
+        except models.ManagedGroup.DoesNotExist:
+            raise Http404("Workspace or ManagedGroup not found.")
+        return group
+
+    def get(self, request, *args, **kwargs):
+        self.group = self.get_group()
+        return super().get(request, *args, **kwargs)
+
+    def post(self, request, *args, **kwargs):
+        self.group = self.get_group()
+        return super().post(request, *args, **kwargs)
+
+    def get_initial(self):
+        initial = super().get_initial()
+        initial["group"] = self.group
+        return initial
+
+    def get_form(self, **kwargs):
+        """Get the form and set the inputs to use a hidden widget."""
+        form = super().get_form(**kwargs)
+        form.fields["group"].widget = HiddenInput()
+        return form
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context["group"] = self.group
+        return context
+
+    def get_success_url(self):
+        return self.object.get_absolute_url()
+
+
+class WorkspaceGroupSharingCreateByWorkspaceGroup(WorkspaceGroupSharingCreate):
+    """View to create a new WorkspaceGroupSharing object for the workspace and group specified in the url."""
+
+    model = models.WorkspaceGroupSharing
+    form_class = forms.WorkspaceGroupSharingForm
+    template_name = (
+        "anvil_consortium_manager/workspacegroupsharing_form_byworkspacegroup.html"
+    )
+    success_msg = "Successfully shared Workspace with Group."
+    """Message to display when the WorkspaceGroupSharing object was successfully created in the app and on AnVIL."""
+
+    message_group_not_found = "Managed Group not found on AnVIL."
+    """Message to display when the ManagedGroup was not found on AnVIL."""
+
+    message_already_exists = (
+        "This workspace has already been shared with this managed group."
+    )
+
+    def get_workspace(self):
+        try:
+            billing_project_slug = self.kwargs["billing_project_slug"]
+            workspace_slug = self.kwargs["workspace_slug"]
+            workspace = models.Workspace.objects.get(
+                billing_project__name=billing_project_slug, name=workspace_slug
+            )
+        except models.Workspace.DoesNotExist:
+            raise Http404("Workspace not found.")
+        return workspace
+
+    def get_group(self):
+        try:
+            group_slug = self.kwargs["group_slug"]
+            group = models.ManagedGroup.objects.get(name=group_slug)
+        except models.ManagedGroup.DoesNotExist:
+            raise Http404("Workspace or ManagedGroup not found.")
+        return group
+
+    def get(self, request, *args, **kwargs):
+        self.workspace = self.get_workspace()
+        self.group = self.get_group()
+        try:
+            obj = models.WorkspaceGroupSharing.objects.get(
+                workspace=self.workspace, group=self.group
+            )
+            messages.error(self.request, self.message_already_exists)
+            return HttpResponseRedirect(obj.get_absolute_url())
+        except models.WorkspaceGroupSharing.DoesNotExist:
+            return super().get(request, *args, **kwargs)
+
+    def post(self, request, *args, **kwargs):
+        self.workspace = self.get_workspace()
+        self.group = self.get_group()
+        try:
+            obj = models.WorkspaceGroupSharing.objects.get(
+                workspace=self.workspace, group=self.group
+            )
+            messages.error(self.request, self.message_already_exists)
+            return HttpResponseRedirect(obj.get_absolute_url())
+        except models.WorkspaceGroupSharing.DoesNotExist:
+            return super().post(request, *args, **kwargs)
+
+    def get_initial(self):
+        initial = super().get_initial()
+        initial["workspace"] = self.workspace
+        initial["group"] = self.group
+        return initial
+
+    def get_form(self, **kwargs):
+        """Get the form and set the inputs to use a hidden widget."""
+        form = super().get_form(**kwargs)
+        form.fields["workspace"].widget = HiddenInput()
+        form.fields["group"].widget = HiddenInput()
+        return form
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context["workspace"] = self.workspace
+        context["group"] = self.group
+        return context
+
+    def get_success_url(self):
+        return self.object.get_absolute_url()
 
 
 class WorkspaceGroupSharingUpdate(